+++ cocos/2d/CCParticleBatchNode.cpp
@@ -401,7 +401,7 @@ void ParticleBatchNode::draw(void)
 
 
 
-void ParticleBatchNode::increaseAtlasCapacityTo(int quantity)
+void ParticleBatchNode::increaseAtlasCapacityTo(ssize_t quantity)
 {
     CCLOG("cocos2d: ParticleBatchNode: resizing TextureAtlas capacity from [%lu] to [%lu].",
           (long)_textureAtlas->getCapacity(),
